+++ otter/assign/__init__.py
@@ -9,8 +9,7 @@ from .assignment import Assignment
 from .output import write_output_directories
 from .utils import run_tests, write_otter_config_file, run_generate_autograder
 
-from ..export import export_notebook
-from ..export.exporters import WkhtmltopdfNotFoundError
+from ..export import export_notebook, WkhtmltopdfNotFoundError
 from ..plugins import PluginCollection
 from ..utils import chdir, get_relpath, knit_rmd_file, loggers