+++ otter/export/utils.py
@@ -1,0 +1,7 @@
+"""Utilities for Otter Export"""
+
+
+class WkhtmltopdfNotFoundError(Exception):
+    """
+    Exception to throw when PDF via HTML is indicated but wkhtmltopdf is not found
+    """